Fix issues with Bool embedding (#207)
* Fix issues in Bool embedding

* Bump MacOS x86_64 image following GitHub deprecation

* fix: don't use `Iff` in the RHS in `Embedding/Bool.lean`

I was assuming the `elimIff` step would get rid of those, but it
appears not to work reliably, so it's better if we avoid `Iff`
entirely.

The order of fvars in the models changed, so I had to update two
tests. We should make that more stable (by sorting fvars?).